==> contrastive/notebooks/archives/ipmi_2023/01-preprocessing_change-keys-in-partial-config.ipynb <==
Notebook to change the keys in partial_config.yaml in all subfolders. It avoids having to rerun the training just to change the keys.

## Change key in partial config

from contrastive.utils.config import create_accessible_config


keys_to_keep = ['dataset_name', 'nb_subjects', 'model', 'with_labels', 'numpy_all',
    'input_size', 'temperature_initial', 'temperature', 'sigma', 'drop_rate', 'depth_decoder',
    'mode', 'foldlabel', 'fill_value', 'patch_size', 'max_angle', 'checkerboard_size', 'keep_bottom',
    'growth_rate', 'block_config', 'num_init_features', 'num_representation_features', 'num_outputs',
    'environment', 'batch_size', 'pin_mem', 'partition', 'lr', 'weight_decay', 'max_epochs',
    'early_stopping_patience', 'seed', 'backbone_name', 'n_max']


main_path = "/neurospin/dico/data/deep_folding/papers/ipmi2023/models/contrastive/trained_on_HCP_half_2/all/no_foldlabel"
dir_paths = os.listdir(main_path)

for dir_path in dir_paths:
    dir_path = main_path +'/'+ dir_path
    for name in os.listdir(dir_path):
        sub_dir = dir_path + '/' + name
        # checks if directory
        if os.path.isdir(sub_dir):
            # check if directory associated to a model
            if os.path.exists(sub_dir+'/.hydra/config.yaml'):
                print("Treating", sub_dir)
                create_accessible_config(keys_to_keep, sub_dir+"/.hydra/config.yaml")
