==> contrastive/notebooks/archives/ipmi_2023/02-training_SimCLR_beta-VAE.ipynb <==
This notebook contains the code to load the latent spaces of beta-VAE and SimCLR

import pandas as pd

src_bVAE = "/neurospin/dico/data/deep_folding/papers/ipmi2023/models/beta-VAE/n_10/#5/cingulate_ACCpatterns_1_embeddings/full_embeddings.csv"
src_SimCLR = "/neurospin/dico/data/deep_folding/papers/ipmi2023/models/contrastive/trained_on_HCP_half_2/all/foldlabel/convnet-10/09-21-24_4/cingulate_ACCpatterns_1_embeddings/full_embeddings.csv"
dst = "/neurospin/dico/data/deep_folding/papers/ipmi2023/models/fusion/#5/cingulate_ACCpatterns_1_embeddings/full_embeddings.csv"


bVAE = pd.read_csv(src_bVAE)
SimCLR = pd.read_csv(src_SimCLR)

fusion = pd.merge(bVAE, SimCLR, left_on='ID', right_on='ID')

fusion.head()

len(fusion)

fusion.set_index('ID', inplace=True)

fusion.head()

fusion.to_csv(dst)
