==> contrastive/notebooks/archives/ipmi_2023/03-analysis_anatomist_visualisation_model_selection.ipynb <==
Notebook to get the extreme predictions' subject name for a given classifier.

import pandas as pd

from contrastive.evaluation.train_multiple_classifiers import compute_auc
from contrastive.utils.models_database import import_bdd

bdd = import_bdd()
bdd

best_model_path = bdd.iloc[0,:].model_path
print(best_model_path)

best_model_class = pd.read_csv(best_model_path+"/cingulate_ACCpatterns_embeddings/cross_val_predicted_labels.csv", index_col=0)
best_model_class

densenet_bdd = bdd[bdd.backbone_name == 'densenet']

best_densenet_path = densenet_bdd.iloc[0,:].model_path
print(best_densenet_path)

best_densenet_class = pd.read_csv(best_densenet_path+"/cingulate_ACCpatterns_embeddings/cross_val_predicted_labels.csv", index_col=0)
best_densenet_class

def get_classifier_by_quality(labels_df, q=0.5):
    aucs = labels_df.apply(compute_auc, args=[labels_df.label])
    aucs = aucs[~aucs.index.isin(['label', 'median_pred', 'mean_pred'])]
    model_auc = aucs.quantile(q=q, interpolation='nearest')
    aucs = aucs[aucs == model_auc]
    return(aucs.index[0], model_auc)

def get_subjects_to_visualize(model_path, q=1):
    model_classifiers = pd.read_csv(model_path+"/cingulate_ACCpatterns_embeddings/cross_val_predicted_labels.csv", index_col=0)

    classifier_name, _ = get_classifier_by_quality(model_classifiers, q=q)

    preds = model_classifiers[['label', classifier_name]]
    preds.sort_values(by=classifier_name, ascending=False, inplace=True)

    subjects_tag = ["worst 1 pred", "best 0 pred", "limit 0 pred",
                    "best 1 pred", "worst 0 pred", "limit 1 pred"]
    subjects_to_visualize = []
    predictions = []

    for i in range(2):
        half_preds = preds[preds.label == i]

        subjects_to_visualize.append(half_preds.index[0])
        subjects_to_visualize.append(half_preds.index[-1])

        predictions.append(half_preds[classifier_name][0])
        predictions.append(half_preds[classifier_name][-1])

        if i == 0:
            half_preds = half_preds[half_preds[classifier_name] <= 0.5]
            if half_preds.values != []:
                subjects_to_visualize.append(half_preds.index[0])
                predictions.append(half_preds[classifier_name][0])

        if i == 1:
            half_preds = half_preds[half_preds[classifier_name] >= 0.5]
            if half_preds.values != []:
                subjects_to_visualize.append(half_preds.index[-1])
                predictions.append(half_preds[classifier_name][-1])

    return pd.DataFrame({'tag': subjects_tag, 'subject':subjects_to_visualize, 'preds': predictions})

subjects_to_visualize = get_subjects_to_visualize("/neurospin/dico/agaudin/Runs/04_pointnet/Output/2022-09-30/17-27-31")
subjects_to_visualize
