==> contrastive/notebooks/archives/ipmi_2023/03-analysis_beta-VAE.ipynb <==
Notebook to preprocess and display classifiers results for the beta-VAEs computed by Louise.

import os
import json
import pandas as pd
import numpy as np

from scipy.stats import ttest_ind

from contrastive.utils.models_database import
 import_bdd

## Get ACCpatterns1 from ACCpatterns embeddings

acc0_subjects = pd.read_csv("/neurospin/dico/data/deep_folding/current/datasets/ACCpatterns/ACCpatterns_1_subject.csv")
acc0_subjects

def get_acc0_embs(embs_path, acc0_embs_path, acc0_subjects=acc0_subjects):
    embs = pd.read_csv(embs_path, index_col=0)
    acc0_embs = embs[embs.index.isin(acc0_subjects.Subject.values)]
    acc0_embs.sort_index(inplace=True)

    # save the acc0 embeddings
    folder_path = '/' + os.path.join(*acc0_embs_path.split('/')[:-1])
    if not os.path.exists(folder_path):
        os.mkdir(folder_path)
    acc0_embs.to_csv(acc0_embs_path)
    return acc0_embs

# for i in range(1,6):
#     embs_path = f"/neurospin/dico/deep_folding/papers/ipmi2023/models/beta-VAE/n_30/#{i}/full_embeddings.csv"
#     acc0_embs_path = '/' + os.path.join(*embs_path.split('/')[:-1]) + '/cingulate_ACCpatterns_1_embeddings/full_embeddings.csv'

#     get_acc0_embs(embs_path, acc0_embs_path)

## Resume beta-VAE results

Visualisation of the latent spaces in latent_space_visualization.ipynb

# compute and plot the dataframe of beta-VAE performances

folder_path = "/neurospin/dico/data/deep_folding/papers/ipmi2023/models/beta-VAE"
dims = [4, 10, 30]
dataset = 'cingulate_ACCpatterns_1'

df = pd.DataFrame(np.zeros((3,1)), index=dims, columns=['beta-VAE'])
std_df = pd.DataFrame(np.zeros((3,1)), index=dims, columns=['beta-VAE'])
bdd = pd.DataFrame(np.zeros((5*len(dims),2)),
                   columns=['num_representation_features', 'auc'])
bdd = bdd.astype({'num_representation_features': int})

for k,dim in enumerate(dims):
    performances = []
    for i in range(1,6):
        values_file = folder_path + f'/n_{dim}/#{i}/{dataset}_embeddings/values.json'
        with open(values_file, 'r') as file:
            values = json.load(file)
            auc = values['cross_val_auc'][0]
            performances.append(auc)
            bdd.loc[5*k+i-1,"auc"] = auc
            bdd.loc[5*k+i-1,"num_representation_features"] = dim
    performances = np.array(performances)
    print(dim, performances)
    df.iloc[k,0] = performances.mean()
    std_df.iloc[k,0] = performances.std()

df

bdd = bdd.astype({'num_representation_features': str})

bdd.head()

bdd.to_csv(f"{folder_path}/bdd_beta-VAE_evaluation-ACCpatterns-1.csv",
            index=False)

std_df

## Comparison with SimCLR convnet

bdd = import_bdd("/neurospin/dico/agaudin/Runs/new_bdd_models_0.csv")
convnets = bdd[(bdd.temperature == 0.1) & (bdd.drop_rate == 0.05) & (bdd.backbone_name == 'convnet')]
convnets

dataset = 'cingulate_ACCpatterns_0'

dims = [4, 10, 30]
foldlabels = [True, False]

for dim in dims:
    # load beta-VAE perf
    beta_VAE_perfs = []
    for i in range(1,6):
        values_file = folder_path + f'/n_{dim}/#{i}/{dataset}_embeddings/values.json'
        with open(values_file, 'r') as file:
            values = json.load(file)
            beta_VAE_perfs.append(values['cross_val_auc'][0])
    beta_VAE_perfs = np.array(beta_VAE_perfs)

    for foldlabel in foldlabels:
        # load SimCLR perfs
        SimCLR_perfs = convnets[(convnets.num_representation_features == dim) & (convnets.foldlabel == foldlabel)].auc

        # compute ttest
        print("Conditions:", dim, foldlabel)
        print("Performances: beta-VAE", beta_VAE_perfs.mean(), "; SimCLR", SimCLR_perfs.mean())
        print(ttest_ind(beta_VAE_perfs, SimCLR_perfs, equal_var=False))
        print("")
