==> contrastive/notebooks/archives/ipmi_2023/10-test_switch-to-numpy.ipynb <==
Test notebook to transition from pickle datasets to numpy.

import os
import pandas as pd
import numpy as np


from contrastive.utils.config import process_config
from contrastive.data.datasets import create_sets_pure_contrastive
from contrastive.data.datamodule import DataModule_PureContrastive

### Configs

pickle_all = "/neurospin/dico/data/deep_folding/current/datasets/hcp/crops/2mm/CINGULATE/mask/Rskeleton.pkl"

numpy_all = "/neurospin/dico/data/deep_folding/current/datasets/hcp/crops/2mm/CINGULATE/mask/Rskeleton.npy"
subjects_all = "/neurospin/dico/data/deep_folding/current/datasets/hcp/crops/2mm/CINGULATE/mask/Rskeleton_subject.csv"
train_val_csv_file = "/neurospin/dico/data/deep_folding/papers/midl2022/HCP_half_1bis.csv"
nb_subjects = -1

## Import the sets

# import pickle data (done only once as it is very long)

pickle_data = pd.read_pickle(pickle_all)
pickle_data.iloc[0,:10]

# rename columns

new_names = []

for name in pickle_data.columns:
    new_names.append(name[-6:])

pickle_data.columns = new_names

pickle_data.columns

### Check if the two initial datasets are the same

def check_same_initial_data(numpy_all, subjects_all, pickle_data, security=100):
    # import the data
    normal_data = np.load(numpy_all, mmap_mode='r')
    normal_subjects = pd.read_csv(subjects_all)

    # check the dimensions
    dims_np = normal_data.shape
    if dims_np[0] != pickle_data.shape[1]:
        raise ValueError("Not the same sample size")
    subj = pickle_data.iloc[0,0]
    if subj.shape != dims_np[1:]:
        raise ValueError("Not the same data dimensions")
    
    # check if same subjects
    common_subjects = normal_subjects[normal_subjects.Subject.isin(list(pickle_data))]
    if len(common_subjects) != len(normal_subjects):
        raise ValueError(f"Difference in the chosen subjects: \
            {len(normal_subjects) - len(common_subjects)} subjects are not the same.")
    
    # check if they are in the same order
    diff = pickle_data.columns.astype(int) != normal_subjects.Subject
    if diff.sum() > 0:
        raise ValueError(f"Subjects are not in the same order (found {diff.sum()}\
    differences).")
    
    # content check
    # only check a random element for time purposes
    for i in range(security):
        rand_idx = np.random.randint(0, len(pickle_data.columns))
        if (pickle_data.iloc[0, rand_idx] != normal_data[rand_idx]).any():
            raise ValueError(f"Content of the two datasets don't match (issue for subject\
        {pickle_data.columns[rand_idx]}).")
    
    return normal_data, normal_subjects

normal_data, normal_subjects = check_same_initial_data(numpy_all, subjects_all, pickle_data)

## Create test and train sets

### Create test and train sets (numpy)

Creates the train and test sets just like in the function create_sets_pure_contrastive

# Gets train_val subjects from csv file
train_val_subjects = pd.read_csv(train_val_csv_file, names=['ID'])
print(f"train_val_subjects = {train_val_subjects}")

# Determines test dataframe
test_subjects = normal_subjects[~normal_subjects.Subject.isin(
    train_val_subjects.ID)].index
len_test = len(test_subjects)
print(f"length of test = {len_test}")
print(f"test_subjects = {test_subjects[:5]}")

# /!\ copy the data to construct test_data
test_data = normal_data[test_subjects]
print(f'test set size: {test_data.shape}')

# Cuts train_val set to requested number
if nb_subjects == -1:
    len_train_val = len(train_val_subjects)
else:
    len_train_val = min(nb_subjects,
                        len(train_val_subjects))
    train_val_subjects = train_val_subjects[:len_train_val]

print(f"length of train/val dataframe: {len_train_val}")

 # Determines train/val dataframe
train_val_subjects_index = normal_subjects[normal_subjects.Subject.isin(
                            train_val_subjects.ID)].index
# /!\ copy the data to construct train_val_data
train_val_data = normal_data[train_val_subjects_index]

### Create test and train sets (pickle)

pickle_normal_subjects = pickle_data.columns.tolist()

# Gets train_val subjects from csv file
train_val_subjects = pd.read_csv(train_val_csv_file, names=['ID']).T
train_val_subjects = train_val_subjects.values[0].tolist()
train_val_subjects = list(map(str, train_val_subjects))
print(f"nb train_val_subjects = {len(train_val_subjects)}")

# Determines test dataframe
pickle_test_subjects = list(set(pickle_normal_subjects).difference(train_val_subjects))
len_test = len(pickle_test_subjects)
print(f"pickle_test_subjects = {len_test}")


pickle_test_data = pickle_data[pickle_data.columns.intersection(pickle_test_subjects)]

# Cuts train_val set to requested number
if nb_subjects == -1:
    len_train_val = len(train_val_subjects)
else:
    len_train_val = min(nb_subjects,
                        len(train_val_subjects))
    train_val_subjects = train_val_subjects[:len_train_val]

print(f"length of train/val dataframe: {len_train_val}")

pickle_train_val_data = pickle_data[pickle_data.columns.intersection(
            train_val_subjects)]

pickle_test_data.iloc[0,:10]

## Check if they are the same

### test_subjects

print(test_subjects)
print(pickle_test_subjects)

pickle_data.columns[test_subjects]

print(pickle_data.columns[test_subjects].to_list() == pickle_test_subjects)
print(set(pickle_data.columns[test_subjects].to_list()) == set(pickle_test_subjects))

Same elements, not inth same order. -> No problem normally.

### test_data

print(len(test_data))
print(len(pickle_test_data.columns))

# exhaustive check

for i in range(len(test_data)):
    print(i, (test_data[i] == pickle_test_data.iloc[0,i]).all())

# only true => ok

### train_val_data

print(len(train_val_data))
print(len(pickle_train_val_data.columns))

# exhaustive check

for i in range(len(train_val_data)):
    print(i, (train_val_data[i] == pickle_train_val_data.iloc[0,i]).all())

# only true => ok

# to be sure it works properly
train_val_data[0,0,0,0,0] += 1

for i in range(len(train_val_data)):
    print(i, (train_val_data[i] == pickle_train_val_data.iloc[0,i]).all())
